+++ src/main/java/gtclassic/common/container/GTContainerTypeFilter.java
@@ -8,8 +8,8 @@ import gtclassic.common.tile.GTTileTypeFilter;
 import ic2.core.IC2;
 import ic2.core.inventory.container.ContainerTileComponent;
 import ic2.core.inventory.gui.GuiIC2;
+import ic2.core.inventory.slots.SlotCustom;
 import ic2.core.inventory.slots.SlotDisplay;
-import ic2.core.inventory.slots.SlotOutput;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.InventoryPlayer;
 import net.minecraft.inventory.ClickType;
@@ -28,7 +28,7 @@ public class GTContainerTypeFilter extends ContainerTileComponent<GTTileTypeFilt
 		this.tile = tile;
 		for (int y = 0; y < 3; ++y) {
 			for (int x = 0; x < 3; ++x) {
-				this.addSlotToContainer(new SlotOutput(player.player, tile, x + y * 3, 98 + x * 18, 5 + y * 18));
+				this.addSlotToContainer(new SlotCustom(tile, x + y * 3, 98 + x * 18, 5 + y * 18, this.tile.filter));
 			}
 		}
 		this.addSlotToContainer(new SlotDisplay(tile, 9, 35, 23));