+++ src/main/java/gtclassic/common/tile/GTTileBufferLarge.java
@@ -81,7 +81,7 @@ public class GTTileBufferLarge extends GTTileBufferBase implements IHasGui {
 	@Override
 	public boolean isInventoryFull() {
 		for (int i = 0; i < 27; ++i) {
-			if (this.inventory.get(i).getCount() < 64) {
+			if (this.inventory.get(i).getCount() < this.inventory.get(i).getMaxStackSize()) {
 				return false;
 			}
 		}