+++ src/main/java/gtclassic/common/util/GTIFilters.java
@@ -7,6 +7,7 @@ import gtclassic.common.GTItems;
 import gtclassic.common.tile.GTTileQuantumChest;
 import gtclassic.common.tile.GTTileQuantumTank;
 import gtclassic.common.tile.GTTileTranslocator;
+import gtclassic.common.tile.GTTileTypeFilter;
 import ic2.api.item.ICustomDamageItem;
 import ic2.api.item.IElectricItem;
 import ic2.core.inventory.filters.IFilter;
@@ -105,6 +106,22 @@ public class GTIFilters {
 			}
 			return noneEmptyStacks == 0; // If all stacks are empty means no filter. If it is more then 1 stack in the
 											// filter then return false because it didnt match the filter.
+		}
+	}
+
+	public static class TypeFilter implements IFilter {
+
+		GTTileTypeFilter tile;
+
+		public TypeFilter(GTTileTypeFilter tile) {
+			this.tile = tile;
+		}
+
+		public boolean matches(ItemStack stack) {
+			if (stack.isEmpty()) {
+				return false;
+			}
+			return GTHelperStack.oreDictStartsWith(stack, this.tile.getCurrentFilter());
 		}
 	}