+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Changed AESU recipe to be more like GT4, less of an endless lapis grind.
 - Added UU Matter Plasma as a dynamite magic fuel!
 - Fixed some recipes not using ender chest ore dict
+- Fixed buffers not checking the correct stack size to see if an inventory was full.
 
 # 1.1.1
 - We wrapping this project up on 1.12.2, no data net, no massive open ended time sink projects. Couple more feature additions and then GTC 1.12.2 is done except bug fixes and QOL.